+++ R/interpolationpoints.R
@@ -122,25 +122,27 @@
                     0.001, 0.25);
 
   df = data.frame(DOY = DOY,
-                  MeanTemperature = as.vector(tmean),
-                  MinTemperature = as.vector(tmin),
-                  MaxTemperature = as.vector(tmax),
-                  Precipitation = as.vector(prec),
-                  MeanRelativeHumidity = rhmean,
-                  MinRelativeHumidity = rhmin,
-                  MaxRelativeHumidity = rhmax,
-                  Radiation = rad,
-                  WindSpeed = Wsp,
-                  WindDirection = Wdp,
-                  PET = pet,
+                  MeanTemperature = as.vector(tmean),#"celsius"),
+                  MinTemperature = as.vector(tmin),#"celsius"),
+                  MaxTemperature = as.vector(tmax),#"celsius"),
+                  Precipitation = as.vector(prec), #"L/m^2"),
+                  MeanRelativeHumidity = rhmean,#"%"),
+                  MinRelativeHumidity = rhmin,#"%"),
+                  MaxRelativeHumidity = rhmax, #"%"),
+                  Radiation = rad,#"MJ"),
+                  WindSpeed = Wsp,#"m/s"),
+                  WindDirection = Wdp,#"degrees"),
+                  PET = pet,#,"L/m^2"),
                   row.names = dates)
   return(df)
 }
 
 interpolationpoints<-function(object, points, dates = NULL,
-                              export=FALSE, exportDir = getwd(), exportFormat = "meteoland/txt",
-                              metadatafile = "MP.txt", verbose=TRUE) {
-  exportFormat = match.arg(exportFormat, c("meteoland/txt", "meteoland/rds", "castanea/txt", "castanea/rds"))
+                              export=FALSE, exportDir = getwd(), exportFile = NULL, 
+                              exportFormat = "meteoland/txt",
+                              metadataFile = "MP.txt", verbose=TRUE) {
+  if(export) exportFormat = match.arg(exportFormat, c("meteoland/txt", "meteoland/rds", "castanea/txt", "castanea/rds", "netCDF"))
+  
   if(!inherits(object,"MeteorologyInterpolationData")) stop("'object' has to be of class 'MeteorologyInterpolationData'.")
   if(!inherits(points,"SpatialPointsTopography")) stop("'points' has to be of class 'SpatialPointsTopography'.")
   intpoints = as(points, "SpatialPoints")
@@ -165,22 +167,33 @@ interpolationpoints<-function(object, points, dates = NULL,
   aspect = points@data$aspect
   longlat = spTransform(points,CRS("+proj=longlat"))
   latitude = longlat@coords[,2]
-
+  bbox = object@bbox
+  
   
   if(exportFormat %in% c("meteoland/txt","castanea/txt")) formatType = "txt"
   else if (exportFormat %in% c("meteoland/rds","castanea/rds")) formatType = "rds"
+  else if (exportFormat %in% c("netCDF")) formatType = "netCDF"
   
   # Define vector of data frames
-  dfvec = vector("list",npoints)
-  dfout = data.frame(dir = rep(exportDir, npoints), filename=paste0(ids,".", formatType))
-  dfout$dir = as.character(dfout$dir)
-  dfout$filename = as.character(dfout$filename)
-  dfout$format = exportFormat
-  rownames(dfout) = ids
-  spdf = SpatialPointsDataFrame(as(points,"SpatialPoints"), dfout)
-  colnames(spdf@coords)<-c("x","y")
-  bbox = object@bbox
   
+  if(export & exportFormat %in% c("meteoland/txt","castanea/txt", "meteoland/rds","castanea/rds")) {
+    dfout = data.frame(dir = rep(exportDir, npoints), filename=paste0(ids,".", formatType))
+    dfout$dir = as.character(dfout$dir)
+    dfout$filename = as.character(dfout$filename)
+    dfout$format = exportFormat
+    rownames(dfout) = ids
+    spdf = SpatialPointsDataFrame(as(points,"SpatialPoints"), dfout)
+    colnames(spdf@coords)<-c("x","y")
+  }
+  else if(export & exportFormat=="netCDF") {
+    if(is.null(exportFile)) stop("File 'exportFile' cannot be null when exporting to netCDF!")
+    ncfile = exportFile
+    if(is.null(dates)) dates = object@dates
+    nc <-.openwritepointNetCDF(coordinates(points), proj4string(points), dates = dates, file = ncfile, overwrite = TRUE, verbose = verbose)
+  }
+  else {
+    dfvec = vector("list",npoints)
+  }
   for(i in 1:npoints) {
     if(verbose) cat(paste("Processing point '",ids[i],"' (",i,"/",npoints,") -",sep=""))
     insidebox = (cc[i,1]>=bbox[1,1] && cc[i,1]<=bbox[1,2]) && (cc[i,2]>=bbox[2,1] && cc[i,2]<=bbox[2,2])
@@ -192,16 +205,25 @@ interpolationpoints<-function(object, points, dates = NULL,
       if(verbose) cat(paste(" done"))
     }
     else {
-      if(dfout$dir[i]!="") f = paste(dfout$dir[i],dfout$filename[i], sep="/")
-      else f = dfout$filename[i]
-      writemeteorologypoint(df, f, exportFormat)
-      if(verbose) cat(paste(" written to ",f, sep=""))
-      if(exportDir!="") f = paste(exportDir,metadatafile, sep="/")
-      else f = metadatafile
-      write.table(as.data.frame(spdf),file= f,sep="\t", quote=FALSE)
+      if(exportFormat!="netCDF") {
+        if(dfout$dir[i]!="") f = paste(dfout$dir[i],dfout$filename[i], sep="/")
+        else f = dfout$filename[i]
+        writemeteorologypoint(df, f, exportFormat)
+        if(verbose) cat(paste(" written to ",f, sep=""))
+        if(exportDir!="") f = paste(exportDir,metadataFile, sep="/")
+        else f = metadataFile
+        write.table(as.data.frame(spdf),file= f,sep="\t", quote=FALSE)
+      } else {
+        if(verbose) cat(paste0(" written to netCDF"))
+        .writemeteorologypointNetCDF(df,nc,i)
+      }
     }
     if(verbose) cat(".\n")
   }
   if(!export) return(SpatialPointsMeteorology(points, dfvec, dates))
-  invisible(spdf)
+  if(exportFormat!="netCDF") {
+    invisible(spdf)
+  } else {
+    .closeNetCDF(ncfile, nc)
+  }
 }