+++ R/writemeteorologypoints.R
@@ -1,3 +1,4 @@
+# Writes one file with meteorology data series of a data frame
 writemeteorologypoint<-function(data, file, format = "meteoland/txt") {
   format = match.arg(format, c("meteoland/txt", "meteoland/rds", "castanea/txt", "castanea/rds"))
   if(format=="castanea/txt" || format=="castanea/rds") {
@@ -20,8 +21,9 @@ writemeteorologypoint<-function(data, file, format = "meteoland/txt") {
     saveRDS(data,file)
   }
 }
+# Writes multiple files, one for each point
 writemeteorologypointfiles<-function(object, dir=getwd(), format ="meteoland/txt",
-                                     metadatafile="MP.txt") {
+                                     metadataFile="MP.txt") {
   format = match.arg(format, c("meteoland/txt", "meteoland/rds", "castanea/txt", "castanea/rds"))
   if(!inherits(object,"SpatialPointsMeteorology")) stop("'object' has to be of class 'SpatialPointsMeteorology'.")
   npoints = length(object@data)
@@ -44,8 +46,22 @@ writemeteorologypointfiles<-function(object, dir=getwd(), format ="meteoland/txt
     else f = dfout$filename[i]
     writemeteorologypoint(object@data[[i]], f, format)
   }
-  if(dir!="") f = paste(dir,metadatafile, sep="/")
-  else f = metadatafile
+  if(dir!="") f = paste(dir,metadataFile, sep="/")
+  else f = metadataFile
   write.table(as.data.frame(spdf),file= f,sep="\t", quote=FALSE)
   invisible(spdf)
+}
+# Writes point meteorology to netCDF
+writemeteorologypoints<-function(object, file, format = "netCDF", add = FALSE, 
+                                 overwrite = FALSE, verbose = FALSE) {
+  if(!inherits(object,"SpatialPointsMeteorology")) stop("'object' has to be of class 'SpatialGridMeteorology'.")
+  if(!add) {
+    nc = .openwritepointNetCDF(object@coords, proj4string = proj4string(object), dates = object@dates,
+                               file=file, overwrite = overwrite, verbose = verbose)
+    .writemeteorologypointsNetCDF(data = object@data, nc=nc, verbose = verbose)
+  } else {
+    nc = .openaddNetCDF(file, verbose=verbose)
+    .addreplacemeteorologypointsNetCDF(object, nc=nc, verbose = verbose)
+  }
+  .closeNetCDF(file,nc, verbose = verbose)
 }